some fixes to make archi compilable on linux

+++ archi/WavefrontObj.hpp
@@ -8,7 +8,7 @@ class WavefrontObj : public Processor
 {
 private:
 	//конвертировать файл из входного потока данных в геометрию
-	ptr<Graphics::EditableGeometry<Vertex, unsigned> > Convert(std::vector<int>& positionNumbers);
+	ptr<Graphics::EditableGeometry<Vertex, size_t> > Convert(std::vector<int>& positionNumbers);
 
 public:
 	String GetCommand() const;
